requirements.txt
fastapi==0.68.0
tortoise-orm==0.16.0
pydantic==1.10.0
uvicorn==0.15.0
jinja2==3.0.2
